#2106 Added ability to automatically choose language for reCAPTCHA
