+++ src/Presentation/Nop.Web.Framework/Mvc/Filters/ValidateCaptchaAttribute.cs
@@ -3,6 +3,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Filters;
 using Microsoft.Extensions.Primitives;
 using Nop.Core.Data;
+using Nop.Core.Domain.Security;
 using Nop.Web.Framework.Security.Captcha;
 
 namespace Nop.Web.Framework.Mvc.Filters
