+++ src/Presentation/Nop.Web/Controllers/NewsController.cs
@@ -5,6 +5,7 @@ using Nop.Core;
 using Nop.Core.Domain.Customers;
 using Nop.Core.Domain.Localization;
 using Nop.Core.Domain.News;
+using Nop.Core.Domain.Security;
 using Nop.Services.Events;
 using Nop.Services.Localization;
 using Nop.Services.Logging;
